refactor: eatinorders.domain 패키지를 하위 패키지 order, ordertable로 구분
