+++ src/main/java/kitchenpos/eatinorders/domain/order/Order.java
@@ -1,9 +1,11 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.order;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+
+import kitchenpos.eatinorders.domain.ordertable.OrderTable;
 
 @Table(name = "orders")
 @Entity
