+++ src/main/java/kitchenpos/eatinorders/domain/order/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.order;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
