+++ src/main/java/kitchenpos/eatinorders/domain/ordertable/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.domain.ordertable;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
